Finish 5-9

+++ bookers/app/controllers/books_controller.rb
@@ -1,9 +1,10 @@
 class BooksController < ApplicationController
   before_action  :authenticate_user!
-  
+
   def index
       @books = Book.all
       @book = Book.new
+      @users = User.all
       @user = current_user
   end
 
@@ -22,19 +23,26 @@ class BooksController < ApplicationController
     # ストロングパラメーターを使用
     @book = Book.find(params[:id])
     @user = current_user
+    @users = User.all
    end
 
   def edit
     @book = Book.find(params[:id])
+    unless @book.user_id == current_user.id then
+      redirect_to books_path
+    end
   end
 
   def destroy
+    @book = Book.find(params[:id])
     # ストロングパラメーターを使用
     book = Book.find(params[:id])
-    # 対象をDBから破棄
-    book.destroy
-    # 一覧画面へリダイレクト
-    redirect_to books_path
+    if @book.user_id == current_user.id then
+      # 対象をDBから破棄
+      book.destroy
+      # 一覧画面へリダイレクト
+      redirect_to books_path
+    end
   end
 
   def update
